test: add integration test to delete a commentary
To delete a commentary it was added the sequence of the logic to perform the exclusion and create the intercept
with the msw for deleting. For the inMemory object mocked, it was update to make a deep copy for security, because
was copying only the array reference (shallow copy), for that reason was added the library Ladash.