test: update deleting comment integration test
To complete this test, it was need to add the verification of when the Toast leave the screen.
In this commit it was added the jest timers mock to track when the animation will finish of, this is
necessary because it has a setTimeout, without it the test will finish but this function will be called
after.